refactor: extract publish and some io
