+++ package.json
@@ -53,6 +53,10 @@
         "@types/node": "^14.14.14",
         "@typescript-eslint/eslint-plugin": "^4.10.0",
         "@typescript-eslint/parser": "^4.10.0",
+        "@yarnpkg/cli": "^2.4.0",
+        "@yarnpkg/core": "^2.4.0",
+        "@yarnpkg/plugin-npm": "^2.4.0",
+        "@yarnpkg/plugin-pack": "^2.2.3",
         "@yarnpkg/pnpify": "^2.4.0",
         "all-contributors-cli": "^6.19.0",
         "babel-jest": "^26.6.3",
@@ -70,6 +74,7 @@
         "jest": "^26.6.3",
         "jest-junit": "^12.0.0",
         "lint-staged": "^10.5.3",
+        "monodeploy-types": "workspace:packages/types",
         "package-json": "^6.5.0",
         "prettier": "^2.2.1",
         "typescript": "^4.1.3",
