+++ packages/io/package.json
@@ -21,5 +21,9 @@
   },
   "devDependencies": {
     "@types/node": "^14.14.25"
+  },
+  "dependencies": {
+    "@yarnpkg/core": "^2.4.0",
+    "@yarnpkg/fslib": "^2.4.0"
   }
 }
