+++ packages/io/src/index.ts
@@ -1,1 +1,2 @@
 export * from './stream'
+export * from './executeLifecycleScript'
