+++ packages/node/src/core/getExplicitVersionStrategies.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import { structUtils } from '@yarnpkg/core'
 import { PortablePath } from '@yarnpkg/fslib'
 
-import { gitDiffTree } from 'monodeploy-git'
+import { getCommitMessages, gitDiffTree } from 'monodeploy-git'
 import logging from 'monodeploy-logging'
 import type {
     MonodeployConfiguration,
@@ -12,7 +12,6 @@ import type {
     YarnContext,
 } from 'monodeploy-types'
 
-import getCommitMessages from '../utils/getCommitMessages'
 import {
     STRATEGY,
     createGetConventionalRecommendedStrategy,
