+++ packages/publish/package.json
@@ -20,6 +20,14 @@
     "prepack": "run workspace:build \"$(pwd)\""
   },
   "dependencies": {
+    "@yarnpkg/cli": "^2.4.0",
+    "@yarnpkg/core": "^2.4.0",
+    "@yarnpkg/fslib": "^2.4.0",
+    "@yarnpkg/plugin-npm": "^2.4.0",
+    "@yarnpkg/plugin-pack": "^2.2.3",
+    "monodeploy-git": "workspace:packages/git",
+    "monodeploy-io": "workspace:packages/io",
+    "monodeploy-logging": "workspace:packages/logging",
     "monodeploy-types": "workspace:packages/types"
   }
 }
